Refactor: Modularize monolithic index.tsx
This refactoring addresses the primary architectural issue of the project: a single, monolithic `index.tsx` file containing all application logic. The codebase has been restructured into a modular architecture within a new `src` directory.

Key changes include:
- Created a `src` directory with subdirectories for `api`, `core`, `features`, `pages`, `shared`, and `styles`.
- Deconstructed `index.tsx` into smaller, single-responsibility modules.
- Created a new main entry point `src/main.ts`.
- Updated `index.html` to point to the new entry point.
- Moved `index.css` to `src/styles/main.css`.
- Cleaned up the root directory by removing the now-obsolete `index.tsx` and `index.css` files.

This new structure significantly improves the maintainability, readability, and scalability of the codebase, making future development and debugging much more efficient.
