+++ index.html
@@ -158,6 +158,6 @@
     <!-- Toast Notification Container -->
     <div id="toast-notification-container" aria-live="assertive" aria-atomic="true"></div>
 
-    <script type="module" src="/index.tsx"></script>
+    <script type="module" src="/src/main.ts"></script>
 </body>
 </html>
