+++ src/pages/index.ts
@@ -1,0 +1,11 @@
+export * from './fisica';
+export * from './mental';
+export * from './financeira';
+export * from './familiar';
+export * from './profissional';
+export * from './social';
+export * from './espiritual';
+export * from './preventiva';
+export * from './planejamento-diario';
+export * from './tarefas';
+export * from './inicio';
