+++ src/styles/main.css
@@ -1,5 +1,3 @@
-
-
 /* General Styles (Base do seu Dash board.html e Saudefisica1.html) */
 :root {
     /* Updated color palette based on user study */
